Enforce score randomization w/o CECP 'random' #3,
as ome CECP interfaces or interfaces pretending to be CECP interfaces
do not act as XBoard / (Winboard) in the regard that after new command
they do not send "random" command as part of initialization string.

Since this 1.0 branch of undead code is meant as fun scholastic engine,
not the boring one to beat with once-discovered winning line(s), it is
desirable to have the slight score randomization in effect at all times.

Completely removing non-deterministic play compile option was not desired,
as deterministic behaviour compiles can be useful for testing purposes.